OWB-548 added null check
git-svn-id: https://svn.apache.org/repos/asf/openwebbeans/trunk@1083864 13f79535-47bb-0310-9956-ffa450edef68

+++ webbeans-impl/src/main/java/org/apache/webbeans/corespi/se/DefaultContextsService.java
@@ -312,8 +312,11 @@ public class DefaultContextsService extends AbstractContextsService
     
     private void stopApplicationContext(Object object)
     {
-        applicationContext.destroy();
-        applicationContext = null;
+        if(applicationContext != null)
+        {
+            applicationContext.destroy();
+            applicationContext = null;
+        }
     }
